axios, routing and auth context improvements
