+++ src/hooks/index.js
@@ -1,5 +1,5 @@
 import { useLocalStorage } from './useLocalStorage';
 import { useAuth } from './useAuth';
-import { useInterceptors } from './useInterceptors'
+import { useAxiosSetup } from './useAxiosSetup';
 
-export { useLocalStorage, useAuth, useInterceptors }
+export { useLocalStorage, useAuth, useAxiosSetup }
